requirements.txt
selenium
webdriver-manager
pandas
sqlalchemy
matplotlib
weasyprint
python-dotenv
smtplib